fix: rename and replace test classes
