+++ src/test/java/ru/practicum/shareit/item/dto/CommentDtoJsonTest.java
@@ -1,0 +1,32 @@
+package ru.practicum.shareit.item.dto;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.json.JsonTest;
+import org.springframework.boot.test.json.JacksonTester;
+import org.springframework.boot.test.json.JsonContent;
+
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+
+@JsonTest
+class CommentDtoJsonTest {
+    @Autowired
+    private JacksonTester<CommentDto> json;
+
+    @Test
+    void testCommentDto() throws Exception {
+        CommentDto commentDto = new CommentDto(
+                1L,
+                "That's the best tool",
+                "Josh",
+                "2023-07-05T15:00:00.000000"
+        );
+
+        JsonContent<CommentDto> result = json.write(commentDto);
+
+        assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(1);
+        assertThat(result).extractingJsonPathStringValue("$.text").isEqualTo("That's the best tool");
+        assertThat(result).extractingJsonPathStringValue("$.authorName").isEqualTo("Josh");
+        assertThat(result).extractingJsonPathStringValue("$.created").isEqualTo("2023-07-05T15:00:00.000000");
+    }
+}
