+++ src/test/java/ru/practicum/shareit/item/dto/CreationCommentDtoJsonTest.java
@@ -1,0 +1,26 @@
+package ru.practicum.shareit.item.dto;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.json.JsonTest;
+import org.springframework.boot.test.json.JacksonTester;
+import org.springframework.boot.test.json.JsonContent;
+
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+
+@JsonTest
+class CreationCommentDtoJsonTest {
+    @Autowired
+    private JacksonTester<CreationCommentDto> json;
+
+    @Test
+    void testCreationCommentDto() throws Exception {
+        CreationCommentDto creationCommentDto = new CreationCommentDto(
+                "That's the best tool"
+        );
+
+        JsonContent<CreationCommentDto> result = json.write(creationCommentDto);
+
+        assertThat(result).extractingJsonPathStringValue("$.text").isEqualTo("That's the best tool");
+    }
+}
