+++ src/test/java/ru/practicum/shareit/item/dto/ItemDtoJsonTest.java
@@ -1,0 +1,44 @@
+package ru.practicum.shareit.item.dto;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.json.JsonTest;
+import org.springframework.boot.test.json.JacksonTester;
+import org.springframework.boot.test.json.JsonContent;
+
+import java.util.Collections;
+
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+
+@JsonTest
+class ItemDtoJsonTest {
+    @Autowired
+    private JacksonTester<ItemDto> json;
+
+    @Test
+    void testItemDto() throws Exception {
+        ItemDto itemDto = new ItemDto(
+                1L,
+                1L,
+                "Hammer",
+                "Handy tool",
+                true,
+                null,
+                null,
+                1L,
+                Collections.emptyList()
+        );
+
+        JsonContent<ItemDto> result = json.write(itemDto);
+
+        assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(1);
+        assertThat(result).extractingJsonPathNumberValue("$.ownerId").isEqualTo(1);
+        assertThat(result).extractingJsonPathStringValue("$.name").isEqualTo("Hammer");
+        assertThat(result).extractingJsonPathStringValue("$.description").isEqualTo("Handy tool");
+        assertThat(result).extractingJsonPathBooleanValue("$.available").isEqualTo(true);
+        assertThat(result).extractingJsonPathValue("$.lastBooking").isEqualTo(null);
+        assertThat(result).extractingJsonPathValue("$.nextBooking").isEqualTo(null);
+        assertThat(result).extractingJsonPathNumberValue("$.requestId").isEqualTo(1);
+        assertThat(result).extractingJsonPathArrayValue("$.comments").isEmpty();
+    }
+}
