--- src/test/java/ru/practicum/shareit/user/UserDtoJsonTest.java
@@ -1,30 +1,0 @@
-package ru.practicum.shareit.user;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.json.JsonTest;
-import org.springframework.boot.test.json.JacksonTester;
-import org.springframework.boot.test.json.JsonContent;
-import ru.practicum.shareit.user.dto.UserDto;
-
-import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-
-@JsonTest
-class UserDtoJsonTest {
-    @Autowired
-    private JacksonTester<UserDto> json;
-
-    @Test
-    void testUserDto() throws Exception {
-        UserDto userDto = new UserDto(
-                1L,
-                "John",
-                "john.doe@mail.com");
-
-        JsonContent<UserDto> result = json.write(userDto);
-
-        assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(1);
-        assertThat(result).extractingJsonPathStringValue("$.name").isEqualTo("John");
-        assertThat(result).extractingJsonPathStringValue("$.email").isEqualTo("john.doe@mail.com");
-    }
-}
